requirements-app.txt
# GitHub App Dependencies

# Web Framework
flask>=3.0.0
gunicorn>=21.2.0

# GitHub API
PyGithub>=2.1.1
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0

# AI Providers
openai>=1.0.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Logging
structlog>=23.2.0
